------------------------------------------------------------------------ r3941 | schka01 | 2010-12-03 13:30:56 -0500 (Fri, 03 Dec 2010) | 7 lines
Prototype and warning fixes for QEF.
Mostly mechanical;  the most interesting discoveries were
some array overruns in qelsel.c, and the improper use of
CMcpychar to copy single-character ascii constants (which is not
correct, as it references past the end of the constant string.)

------------------------------------------------------------------------
r3942 | wanfr01 | 2010-12-03 14:00:18 -0500 (Fri, 03 Dec 2010) | 59 lines

Review Board IP  http://reviewboard.ingres.prv/r/7018

Cross integration of change 508505

(REPOST to add missing cbf changes)

SIR 124714 - Add a new parameter to tune disk I/O QEP estimates.

Issue 144340 deals with a query whose query plan changed between 2.6 and 9.2.
On one particular machine (a machine that appears to have a slow disk), the new
query plan is running 8 times slower.  The actual runtime of the query is very
small, but it is run extremely frequently.

Post 2.6, a few changes were made which changed the costs of sorts and
disk I/O. This could result in unexpected query plan changes which may or may
not be appropriate.

For part of that change, dbms parameters were created which allow you to adjust
the modifiers of the sorts.  These two undocumented parameters, sort_cpu_factor
and sort_io_factor, were added in change 482555.

Change 484482 modified opn_sblock (which is derived from ops_scanblocks), which
was the estimated number of blocks per IO in a hold file.  Change 484482
doubled the value which in theory corrected for the underestimation originally
set by change 320002.

The result of that change, for this particular test case, is that a full
sort merge QEP cost estimate is now less than the original K-join QEP.
On this particular machine, performance takes a dive, most likely due to the
disk.

This change adds a new dbms parameter, block_holdfactor, that allows you to
modify the adjustment to opn_sblock.  The default, 100, leaves the adjustment
at current 9.2 levels.  Setting the value to 200 will set the adjustment back to
2.6 levels. I am using sort_io_factor as a model and am using similar
datatypes.

In theory, this value should be increased to compensate for slower disks.
The default values appear to be fine for most machines.

Side note: help file to be added in separate IP since my psuite is not working
for 'new files'

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------