------------------------------------------------------------------------ r3956 | hanje04 | 2010-12-06 14:00:12 -0500 (Mon, 06 Dec 2010) | 4 lines
Add back modifyfe for VMS after the LSB changed removed it from the
dbsupport package.

------------------------------------------------------------------------
r3957 | thaju02 | 2010-12-06 15:19:42 -0500 (Mon, 06 Dec 2010) | 89 lines

Review Board IP  http://reviewboard.ingres.prv/r/7085

Cross integration of change 506902

Bug 124252: Preparing a long insert statement after preparing and
running a long select statement with many unicode parameters fails
with E_AD2004_BAD_DTID and E_SC021C_SCS_INPUT_ERROR.

The E_AD2004 stems from adi_dtinfo() called by scs_fetch_data()
during the "prepare JDBC_STMT_0_1 into sqlda from INSERT INTO...".
The DAS sends the prepare stmt to the dbms across 7 gca consecutive
buffers; the first 6 buffers are of length 4064 bytes and the last
buffer is of length 2468 bytes. Upon receipt of the last buffer,
the dbms scs_fetch_data()/scs_qt_fetch() parses the first 2456 bytes
of DB_QTXT_TYPE data. scs_fetch_data() then expects the remaining
12 bytes to be GCA_DATA_VALUE struct data, but instead the remaining
12 bytes are null. scs_fetch_data() unknowingly calls adi_dtinfo()
with gca_type of 0 and adi_dtinfo() in turn returns E_AD2004 due
to the invalid gca_type.

Excerpts from communications with Gordy on DAS:
"Src and src1 do get redirected during gco_encode() processing
due to the need to follow pointers in the input data structures.
This is normal, looks strange, but otherwise does not cause a
problem. Seeing usrlen go negative usually doesn't cause a
problem. At the start of processing, usrlen is used to setup
src2. In the midst of processing, when src1 is redirected,
usrlen is changed at the bottom of SD_FMT to allow changes in
src1/src2 to be tracked and may be very large or go negative.
Upon re-entry to SD_FMT, that change is basically reversed so
that src1/src2 should be the same as at the end of the previous
instance. At the end of processing, src1 will be restored to
the proper context and usrlen will then make more sense.

It is the bad test result in OP_JNINP which leads directly to
the 12 bytes being added to the buffer.  The twelve bytes are
the first three elements of the phantom GCA_DATA_VALUE following
the query text.  Once those three elements are processed,
gco_encode() attempts to decipher the data type for processing
and hits the same sort of error the DBMS is seeing and processing
of the message stops.  The result is sent as is to the DBMS...

I think the problem IS 64 bit related. On 32 bit systems, usrlen
can properly hold the difference between src1 and src2.  Even if
usrlen is negative, src2 will get reset to proper position.
Usrlen is an i4, so on 64 bit systems may not be able to hold
the difference between src1 and src2 when src1 has been redirected
to a different context.  This results in the upper bits of src2
getting set incorrectly.

A quick fix would be to change usrlen to an i8.  That should fix
the problem, but it is not the ultimate clean solution.  Probably
the correct thing to do is to make sure that in gco_encode() and
gco_decode() that src, src1 and src2 are always in the same
context. If src is changed to follow a pointer, then src1/src2
should probably be changed to the same context (and similarily
when src is restored).  The other option is to make sure both
src1 and src2 stay in the original context and that neither get
changed when src follows a pointer.  This would require a thorough
check of all processing actions and additional storage in the DOC
structure to track the correct values...

In ingres10, API no longer uses the formatted GCA interface
(deprecating that option).  I'm planning on removing support
altogether for the formatted interface after ingres10.

This problem is due to unusual values returned from gco_encode()
during message processing.  This will be remove when I get rid
of the GCA formatted interface."

This change defines GCA_SR_DM's usrlen as i8 for 64 bit platforms
and i4 for 32 bit.

Added Functions:

   None.

Removed Functions:

   None.

CL Interface changes:

   None.

------------------------------------------------------------------------