------------------------------------------------------------------------ r3965 | horda03 | 2010-12-08 04:00:16 -0500 (Wed, 08 Dec 2010) | 16 lines
Review Board IP  http://reviewboard.ingres.prv/r/7101

Fix VMS build problems in MAIN on VMS introduced by the GESO Spatial project.

The contents of a LOCATION are not accessible to code outside of the CL (these days
with Unix and VMS sharing code - not a good idea in CL either) as the contents
will differ depending on the OS. ducommon.qsc was accessing "path" replaced with
LOtos().

There is a VMS limitation on objects in a library. They must be unique within the
first 32 characters, adu_geometrycollection_srid_fromWKB and
adu_geometrycollection_srid_fromText are not unique withinthe first 32 characters,
so renamed to adu_geomcollection_srid_fromWKB and adu_geomcollection_srid_fromText
respectively.

------------------------------------------------------------------------
r3966 | hanal04 | 2010-12-08 04:30:35 -0500 (Wed, 08 Dec 2010) | 11 lines

Cross integration of change 508841

Review Board IP  http://reviewboard.ingres.prv/r/7011

If a report generates no rows and is not forced to the printer
we were failing to delete the $II_TEMPORARY/raxxxxx.tmp file
created to hold the report content. This change adds that
cleanup.

------------------------------------------------------------------------